фикс транзакции
